Forced upgrade 1.7.9

+++ TDLib.Api/Functions/AddChatMember.cs
@@ -10,7 +10,7 @@ namespace TdLib
     public static partial class TdApi
     {
         /// <summary>
-        /// Adds a new member to a chat. Members can't be added to private or secret chats. Members will not be added until the chat state has been synchronized with the server
+        /// Adds a new member to a chat. Members can't be added to private or secret chats
         /// </summary>
         public class AddChatMember : Function<Ok>
         {
@@ -38,10 +38,10 @@ namespace TdLib
             /// </summary>
             [JsonConverter(typeof(Converter))]
             [JsonProperty("user_id")]
-            public int UserId { get; set; }
+            public long UserId { get; set; }
 
             /// <summary>
-            /// The number of earlier messages from the chat to be forwarded to the new member; up to 100. Ignored for supergroups and channels
+            /// The number of earlier messages from the chat to be forwarded to the new member; up to 100. Ignored for supergroups and channels, or if the added user is a bot
             /// </summary>
             [JsonConverter(typeof(Converter))]
             [JsonProperty("forward_limit")]
@@ -49,10 +49,10 @@ namespace TdLib
         }
 
         /// <summary>
-        /// Adds a new member to a chat. Members can't be added to private or secret chats. Members will not be added until the chat state has been synchronized with the server
+        /// Adds a new member to a chat. Members can't be added to private or secret chats
         /// </summary>
         public static Task<Ok> AddChatMemberAsync(
-            this Client client, long chatId = default, int userId = default, int forwardLimit = default)
+            this Client client, long chatId = default, long userId = default, int forwardLimit = default)
         {
             return client.ExecuteAsync(new AddChatMember
             {
