+++ TDLib.Api/Functions/GetLoginUrl.cs
@@ -45,7 +45,7 @@ namespace TdLib
             /// </summary>
             [JsonConverter(typeof(Converter))]
             [JsonProperty("button_id")]
-            public int ButtonId { get; set; }
+            public long ButtonId { get; set; }
 
             /// <summary>
             /// True, if the user allowed the bot to send them messages
@@ -59,7 +59,7 @@ namespace TdLib
         /// Returns an HTTP URL which can be used to automatically authorize the user on a website after clicking an inline button of type inlineKeyboardButtonTypeLoginUrl.
         /// </summary>
         public static Task<HttpUrl> GetLoginUrlAsync(
-            this Client client, long chatId = default, long messageId = default, int buttonId = default,
+            this Client client, long chatId = default, long messageId = default, long buttonId = default,
             bool allowWriteAccess = default)
         {
             return client.ExecuteAsync(new GetLoginUrl
