+++ TDLib.Api/Functions/GetLoginUrlInfo.cs
@@ -45,14 +45,14 @@ namespace TdLib
             /// </summary>
             [JsonConverter(typeof(Converter))]
             [JsonProperty("button_id")]
-            public int ButtonId { get; set; }
+            public long ButtonId { get; set; }
         }
 
         /// <summary>
         /// Returns information about a button of type inlineKeyboardButtonTypeLoginUrl. The method needs to be called when the user presses the button
         /// </summary>
         public static Task<LoginUrlInfo> GetLoginUrlInfoAsync(
-            this Client client, long chatId = default, long messageId = default, int buttonId = default)
+            this Client client, long chatId = default, long messageId = default, long buttonId = default)
         {
             return client.ExecuteAsync(new GetLoginUrlInfo
             {
