+++ TDLib.Api/Functions/TransferChatOwnership.cs
@@ -38,7 +38,7 @@ namespace TdLib
             /// </summary>
             [JsonConverter(typeof(Converter))]
             [JsonProperty("user_id")]
-            public int UserId { get; set; }
+            public long UserId { get; set; }
 
             /// <summary>
             /// The password of the current user
@@ -52,7 +52,7 @@ namespace TdLib
         /// Changes the owner of a chat. The current user must be a current owner of the chat. Use the method canTransferOwnership to check whether the ownership can be transferred from the current session. Available only for supergroups and channel chats
         /// </summary>
         public static Task<Ok> TransferChatOwnershipAsync(
-            this Client client, long chatId = default, int userId = default, string password = default)
+            this Client client, long chatId = default, long userId = default, string password = default)
         {
             return client.ExecuteAsync(new TransferChatOwnership
             {
