+++ TDLib.Api/Objects/MessageChatUpgradeFrom.cs
@@ -39,7 +39,7 @@ namespace TdLib
                 /// </summary>
                 [JsonConverter(typeof(Converter))]
                 [JsonProperty("basic_group_id")]
-                public int BasicGroupId { get; set; }
+                public long BasicGroupId { get; set; }
             }
         }
     }
