+++ TDLib.Api/Objects/Poll.cs
@@ -33,7 +33,7 @@ namespace TdLib
             public long Id { get; set; }
 
             /// <summary>
-            /// Poll question, 1-300 characters
+            /// Poll question; 1-300 characters
             /// </summary>
             [JsonConverter(typeof(Converter))]
             [JsonProperty("question")]
@@ -58,7 +58,7 @@ namespace TdLib
             /// </summary>
             [JsonConverter(typeof(Converter))]
             [JsonProperty("recent_voter_user_ids")]
-            public int[] RecentVoterUserIds { get; set; }
+            public long[] RecentVoterUserIds { get; set; }
 
             /// <summary>
             /// True, if the poll is anonymous
